Add console script to setup.py
With this change, it is possible to run the server by just running `timetagger`, as an alternative to `python -m timetagger`.

Move timetagger/__main__.py to timetagger/_serve.py and create a new __main__.py for special handling of --version.
